\name{DataCubeReader-package}
\alias{DataCubeReader-package}
\alias{DataCubeReader}
\docType{package}
\title{Reads data from the Earth System Datacube}
}
\description{This is the official R interface for reading data from the Earth System data cube.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DataCubeReader}
\packageIndices{DataCubeReader}
}
\author{Fabian Gans}

Maintainer: \packageMaintainer{DataCubeReader}
}

\keyword{ package }


