\name{open.cube}
\alias{open.cube}
\title{Open an Earth System Data Cube}
\description{Open the Datacube by scanning datacube's base directory and returns an opaque DataCube object handle}
\usage{\method{open}{cube}(cubepath)}
\arguments{
  \item{cubepath}{Location the the DataCube}
}

\value{A list a list containing some cube info, to be used for reading cube data.}

\author{Fabian Gans, Miguel D. Mahecha, MPI BGC Jena, Germany, fgans@bgc-jena.mpg.de mmahecha@bgc-jena.mpg.de
Fabian Gans}





